added single consolidated function and deleted old version, updated name of sea ice vol flux file in native grid datasets
