+++ metadata/ECCOv4r4_metadata_json/ECCOv4r4_groupings_for_native_datasets.json
@@ -68,7 +68,7 @@
     "name": "sea-ice and snow horizontal volume fluxes",
     "fields":"ADVxHEFF, ADVyHEFF, ADVxSNOW , ADVySNOW, DFxESNOW, DFyEHEFF, DFxEHEFF, DFyESNOW",
     "product": "native",
-    "filename" : "SEA_ICE_VOLUME_FLUX",
+    "filename" : "SEA_ICE_HORIZ_VOLUME_FLUX",
     "dimension" : "2D"
    },
    {
